server/requirements.txt
Flask==3.0.0
google-cloud-datastore==2.19.0
PyJWT==2.8.0
bcrypt==4.1.2
python-dotenv==1.0.0
gunicorn==21.2.0
pytz==2024.1
pytest==7.4.3
pytest-flask==1.3.0
